Extraction percent reflects work done, not file position
The Cues index sits at the end of the file, so the raw byte position
jumped to 99% instantly and sat there while the real work (visiting the
indexed blocks) ran. Report indexed-cues-visited on the fast path and
position-over-file-size on the scan path instead — the number now climbs
honestly through a long network extraction.

+++ app/src/main/java/com/newash/videocast/subs/EmbeddedSubtitles.kt
@@ -54,14 +54,14 @@ object EmbeddedSubtitles {
     /**
      * Extracts the track and renders it as WebVTT. [onOpen] hands out the live
      * stream so the caller can close it to abort a blocked read on cancel;
-     * [onProgress] reports the byte position while parsing (MKV only).
+     * [onProgress] reports work done in percent (MKV only).
      */
     fun extractVtt(
         context: Context,
         uri: Uri,
         track: Track,
         onOpen: (AutoCloseable) -> Unit = {},
-        onProgress: (Long) -> Unit = {},
+        onProgress: (Int) -> Unit = {},
     ): String {
         val cues = when (track.container) {
             // A real file descriptor unlocks the walker's Cues-index fast path
